+++ tests/ote_cli/misc/test_ote_cli_common.py
