+++ tests/ote_cli/misc/test_ote_cli_demo.py
@@ -21,12 +21,13 @@ import pytest
 from ote_sdk.test_suite.e2e_test_system import e2e_pytest_component
 from ote_cli.registry import Registry
 
-from common import (
+from test_ote_cli_common import (
     create_venv,
     get_some_vars,
     wrong_paths,
     ote_common,
-    logger
+    logger,
+    default_train_args_paths
 )
 
 
@@ -56,6 +57,29 @@ class TestDemoCommon:
     def create_venv_fx(self, template):
         work_dir, template_work_dir, algo_backend_dir = get_some_vars(template, root)
         create_venv(algo_backend_dir, work_dir, template_work_dir)
+
+    @pytest.fixture()
+    @e2e_pytest_component
+    @pytest.mark.parametrize("back_end, template", params_values, ids=params_ids)
+    def get_pretrained_artifacts_fx(self, template, create_venv_fx):
+        _, template_work_dir, _ = get_some_vars(template, root)
+        pretrained_artifact_path = f"{template_work_dir}/trained_{template.model_template_id}"
+        command_args = [
+            template.model_template_id,
+            "--train-ann-file",
+            f'{os.path.join(ote_dir, default_train_args_paths["--train-ann-file"])}',
+            "--train-data-roots",
+            f'{os.path.join(ote_dir, default_train_args_paths["--train-data-roots"])}',
+            "--val-ann-file",
+            f'{os.path.join(ote_dir, default_train_args_paths["--val-ann-file"])}',
+            "--val-data-roots",
+            f'{os.path.join(ote_dir, default_train_args_paths["--val-data-roots"])}',
+            "--save-model-to",
+            pretrained_artifact_path,
+        ]
+        command_args.extend(default_train_args_paths["train_params"])
+        ote_common(template, root, 'train', command_args)
+        assert os.path.exists(pretrained_artifact_path), f"test artifact must exits by path {pretrained_artifact_path}"
 
     @e2e_pytest_component
     @pytest.mark.parametrize("back_end, template", params_values, ids=params_ids)
@@ -155,3 +179,95 @@ class TestDemoCommon:
         ret = ote_common(template, root, 'demo', command_args)
         assert ret['exit_code'] != 0, "Exit code must not be equal 0"
         assert error_string in ret['stderr'], f"Different error message {ret['stderr']}"
+
+    @e2e_pytest_component
+    @pytest.mark.parametrize("back_end, template", params_values, ids=params_ids)
+    def test_ote_demo_fit_size(self, back_end, template, create_venv_fx, get_pretrained_artifacts_fx):
+        _, template_work_dir, _ = get_some_vars(template, root)
+        command_args = [template.model_template_id,
+                        '--load-weights',
+                        f'{template_work_dir}/trained_{template.model_template_id}/weights.pth',
+                        '--input',
+                        '--delay',
+                        '-1',
+                        f'{os.path.join(ote_dir, "data/airport/train")}',
+                        '--fit-to-size', '1', '1']
+        ret = ote_common(template, root, 'demo', command_args)
+        assert ret['exit_code'] != 0, "Exit code must not be equal 0"
+
+    @e2e_pytest_component
+    @pytest.mark.parametrize("back_end, template", params_values, ids=params_ids)
+    def test_ote_demo_delay_wrong_type(self, back_end, template, create_venv_fx):
+        error_string = "invalid int value"
+        command_args = [template.model_template_id,
+                        '--load-weights',
+                        f'./trained_{template.model_template_id}/weights.pth',
+                        '--input',
+                        f'{os.path.join(ote_dir, "data/airport/train")}',
+                        '--delay',
+                        'String']
+        ret = ote_common(template, root, 'demo', command_args)
+        assert ret['exit_code'] != 0, "Exit code must not be equal 0"
+        assert error_string in ret['stderr'], f"Different error message {ret['stderr']}"
+
+    @e2e_pytest_component
+    @pytest.mark.parametrize("back_end, template", params_values, ids=params_ids)
+    def test_ote_demo_loop_wrong_type(self, back_end, template, create_venv_fx):
+        error_string = "Boolean value expected"
+        command_args = [template.model_template_id,
+                        '--load-weights',
+                        f'./trained_{template.model_template_id}/weights.pth',
+                        '--input',
+                        f'{os.path.join(ote_dir, "data/airport/train")}',
+                        '--loop',
+                        'NotBoolean']
+        ret = ote_common(template, root, 'demo', command_args)
+        assert ret['exit_code'] != 0, "Exit code must not be equal 0"
+        assert error_string in ret['stderr'], f"Different error message {ret['stderr']}"
+
+    @e2e_pytest_component
+    @pytest.mark.parametrize("back_end, template", params_values, ids=params_ids)
+    def test_ote_demo_loop(self, back_end, template, create_venv_fx, get_pretrained_artifacts_fx):
+        _, template_work_dir, _ = get_some_vars(template, root)
+        command_args = [template.model_template_id,
+                        '--load-weights',
+                        f'{template_work_dir}/trained_{template.model_template_id}/weights.pth',
+                        '--input',
+                        f'{os.path.join(ote_dir, "data/airport/train")}',
+                        '--delay',
+                        '-1',
+                        '--loop',
+                        'False']
+        ret = ote_common(template, root, 'demo', command_args)
+        assert ret['exit_code'] == 0, "Exit code must not equal 0"
+
+    @e2e_pytest_component
+    @pytest.mark.parametrize("back_end, template", params_values, ids=params_ids)
+    def test_ote_demo_display_perf_wrong_type(self, back_end, template, create_venv_fx):
+        error_string = "Boolean value expected"
+        command_args = [template.model_template_id,
+                        '--load-weights',
+                        f'./trained_{template.model_template_id}/weights.pth',
+                        '--input',
+                        f'{os.path.join(ote_dir, "data/airport/train")}',
+                        '--display-perf',
+                        'NotBoolean']
+        ret = ote_common(template, root, 'demo', command_args)
+        assert ret['exit_code'] != 0, "Exit code must not be equal 0"
+        assert error_string in ret['stderr'], f"Different error message {ret['stderr']}"
+
+    @e2e_pytest_component
+    @pytest.mark.parametrize("back_end, template", params_values, ids=params_ids)
+    def test_ote_demo_display_perf(self, back_end, template, create_venv_fx, get_pretrained_artifacts_fx):
+        _, template_work_dir, _ = get_some_vars(template, root)
+        command_args = [template.model_template_id,
+                        '--load-weights',
+                        f'{template_work_dir}/trained_{template.model_template_id}/weights.pth',
+                        '--input',
+                        f'{os.path.join(ote_dir, "data/airport/train")}',
+                        '--delay',
+                        '-1',
+                        '--display-perf',
+                        'False']
+        ret = ote_common(template, root, 'demo', command_args)
+        assert ret['exit_code'] == 0, "Exit code must not equal 0"
